lab_1
This dataset explores medical students' empathy, mental health, and burnout in Switzerland. It compiles important demographic information as well as self-reported data and results from psychological tests to give a comprehensive picture of the mental states of students in the medical field. Through this research, we hope to gain a better understanding of how being a medical student can affect health and wellbeing. By looking at the individual factors that may contribute to different outcomes, we can work towards improving educational systems for the benefit of both students and their eventual patients